clean up loading a little
